started working on burger time. level loading, walking around, burger parts falling

+++ BurgerTime/HealthComponent.cpp
@@ -1,4 +1,5 @@
 #include "HealthComponent.h"
+#include "Events.h"
 
 HealthComponent::HealthComponent(dae::GameObject* owner, int amountOfLives)
 	: Component(owner)
@@ -12,5 +13,5 @@ void HealthComponent::Damage(int amount)
 	if (m_AmountOfLives < 0)
 		m_AmountOfLives = 0;
 
-	Notify(this, Events::playerLostLife);
+	Notify(this, static_cast<int>(Events::playerLostLife));
 }
