+++ BurgerTime/PointsComponent.cpp
@@ -1,4 +1,5 @@
 #include "PointsComponent.h"
+#include "Events.h"
 
 PointsComponent::PointsComponent(dae::GameObject* owner)
 	: Component(owner)
@@ -8,5 +9,5 @@ void PointsComponent::AddPoints(int amount)
 {
 	m_AmountOfPoints += amount;
 
-	Notify(this, Events::playerEarnedPoint);
+	Notify(this, static_cast<int>(Events::playerEarnedPoint));
 }
